refactor component types into engine
