+++ src/systems/fov.system.ts
@@ -1,5 +1,10 @@
-import { getQuery, addComponent, getEntity, removeComponent } from "../engine";
-import { ComponentTypes } from "../engine/index.types";
+import {
+  ComponentTypes,
+  getQuery,
+  addComponent,
+  getEntity,
+  removeComponent,
+} from "../engine";
 import { getState } from "../main";
 import createFOV from "../lib/fov";
 import { QueryTypes } from "../queries";
@@ -8,17 +13,17 @@ export const fovSystem = () => {
   const inFovQuery = getQuery(QueryTypes.IsInFov);
   const opaqueQuery = getQuery(QueryTypes.IsOpaque);
 
-  const playerEntity = getEntity(getState().playerEId)
+  const playerEntity = getEntity(getState().playerEId);
 
   if (!playerEntity) {
-    console.log('no player entity')
-    return
+    console.log("no player entity");
+    return;
   }
 
-  const origin = playerEntity.components.position
+  const origin = playerEntity.components.position;
 
   if (!origin) {
-    return
+    return;
   }
 
   const FOV = createFOV(
